provider/aws: combine aws alias tests
