+++ builtin/providers/aws/resource_aws_iam_account_alias_test.go
@@ -3,6 +3,7 @@ package aws
 import (
 	"fmt"
 	"testing"
+	"time"
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/iam"
@@ -12,6 +13,29 @@ import (
 )
 
 func TestAccAWSIAMAccountAlias_basic(t *testing.T) {
+	testCases := map[string]map[string]func(t *testing.T){
+		"Basic": {
+			"basic": testAccAWSIAMAccountAlias_basic_with_datasource,
+		},
+		// "Import": {
+		// 	"import": testAccAWSIAMAccountAlias_importBasic,
+		// },
+	}
+
+	for group, m := range testCases {
+		m := m
+		t.Run(group, func(t *testing.T) {
+			for name, tc := range m {
+				tc := tc
+				t.Run(name, func(t *testing.T) {
+					tc(t)
+				})
+			}
+		})
+	}
+}
+
+func testAccAWSIAMAccountAlias_basic_with_datasource(t *testing.T) {
 	var account_alias string
 
 	rstring := acctest.RandString(5)
@@ -22,9 +46,15 @@ func TestAccAWSIAMAccountAlias_basic(t *testing.T) {
 		CheckDestroy: testAccCheckAWSIAMAccountAliasDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccAWSIAMAccountAliasConfig(rstring),
+				Config: testAccAWSIAMAccountAliasConfig_alias_only(rstring),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckAWSIAMAccountAliasExists("aws_iam_account_alias.test", &account_alias),
+				),
+			},
+			{
+				Config: testAccAWSIAMAccountAliasConfig(rstring),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckAwsIamAccountAlias("data.aws_iam_account_alias.current"),
 				),
 			},
 		},
@@ -78,6 +108,23 @@ func testAccCheckAWSIAMAccountAliasExists(n string, a *string) resource.TestChec
 
 		*a = aws.StringValue(resp.AccountAliases[0])
 
+		time.Sleep(10 * time.Second)
+
+		return nil
+	}
+}
+
+func testAccCheckAwsIamAccountAlias(n string) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+		rs, ok := s.RootModule().Resources[n]
+		if !ok {
+			return fmt.Errorf("Can't find Account Alias resource: %s", n)
+		}
+
+		if rs.Primary.Attributes["account_alias"] == "" {
+			return fmt.Errorf("Missing Account Alias")
+		}
+
 		return nil
 	}
 }
@@ -87,5 +134,13 @@ func testAccAWSIAMAccountAliasConfig(rstring string) string {
 resource "aws_iam_account_alias" "test" {
   account_alias = "terraform-%s-alias"
 }
-`, rstring)
+
+data "aws_iam_account_alias" "current" {}`, rstring)
+}
+
+func testAccAWSIAMAccountAliasConfig_alias_only(rstring string) string {
+	return fmt.Sprintf(`
+resource "aws_iam_account_alias" "test" {
+  account_alias = "terraform-%s-alias"
+}`, rstring)
 }
